Singleton approach to persist known keys across an instance
This way, checking the file system in the exception_keys method only happens once, per instance, when the instance is initialized.

It does not hit the file system every single time handle_error is called.
