+++ lib/vet360/stats.rb
@@ -6,19 +6,6 @@ module Vet360
     FINAL_FAILURE = %w[REJECTED COMPLETED_FAILURE].freeze
 
     class << self
-      # Parses our exceptions file and returns all of the Vet360 exception keys.
-      #
-      # @return [Array] An array of lowercased, alphabetized, Vet360 exception keys
-      #
-      def exception_keys
-        exceptions_file
-          .dig('en', 'common', 'exceptions')
-          .keys
-          .select { |exception| exception.include? 'VET360_' }
-          .sort
-          .map(&:downcase)
-      end
-
       # Triggers the associated StatsD.increment method for the Vet360 buckets that are
       # initialized in the config/initializers/statsd.rb file.
       #
@@ -49,12 +36,6 @@ module Vet360
       end
 
       private
-
-      def exceptions_file
-        config = Rails.root + 'config/locales/exceptions.en.yml'
-
-        YAML.load_file(config)
-      end
 
       def status_in(response)
         response&.body&.dig('tx_status')&.upcase
