+++ spec/lib/vet360/stats_spec.rb
@@ -5,21 +5,6 @@ require 'rails_helper'
 describe Vet360::Stats do
   let(:statsd_prefix) { Vet360::Service::STATSD_KEY_PREFIX }
 
-  describe '.exception_keys' do
-    subject { described_class.exception_keys }
-
-    it 'returns an array of Vet360 exception keys' do
-      expect(subject).to be_a Array
-    end
-
-    it 'contains only downcased, Vet360 exception keys' do
-      total_key_count  = subject.size
-      keys_with_vet360 = subject.select { |key| key.include? 'vet360_' }.size
-
-      expect(total_key_count).to eq keys_with_vet360
-    end
-  end
-
   describe '.increment' do
     it 'increments the StatsD Vet360 counter' do
       bucket1 = 'exceptions'
